+++ mln/labeling/rag.hpp
@@ -134,7 +134,7 @@ namespace mln
 
             if (removed) {
 	      tomerge.push_back(v);
-	      int n = tomerge.size();
+	      int n = (int)tomerge.size();
 	      for (int i = 0; i < n; ++i)
 		for (int j = i+1; j < n; ++j)
 		  if (tomerge[i] != tomerge[j])
